requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.2
celery==5.5.3
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.0.0
Django==5.2.7
django-autoslug==1.9.9
django-celery-beat==2.8.1
django-redis==6.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
kombu==5.5.4
msgpack==1.1.2
packaging==25.0
Pillow==10.1.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
django-environ==0.11.2
redis==5.0.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.0.1